refactor: extract core features
